+++ src/enums/node-transformers/NodeTransformationStage.ts
@@ -1,9 +1,9 @@
 export enum NodeTransformationStage {
     Initializing = 'Initializing',
     Preparing = 'Preparing',
-    MangleProperties = 'MangleProperties',
     DeadCodeInjection = 'DeadCodeInjection',
     ControlFlowFlattening = 'ControlFlowFlattening',
+    RenameProperties = 'RenameProperties',
     Converting = 'Converting',
     Obfuscating = 'Obfuscating',
     Finalizing = 'Finalizing'